refactor application package
