+++ src/main/kotlin/org/ionproject/integration/application/dispatcher/IDispatcher.kt
@@ -1,0 +1,13 @@
+package org.ionproject.integration.application.dispatcher
+
+import org.ionproject.integration.application.dto.ParsedData
+import org.ionproject.integration.infrastructure.OutputFormat
+
+interface IDispatcher {
+    fun dispatch(data: ParsedData, filename: String, format: OutputFormat): DispatchResult
+}
+
+enum class DispatchResult {
+    SUCCESS,
+    FAILURE
+}
