+++ src/main/kotlin/org/ionproject/integration/application/dto/InstitutionMetadata.kt
@@ -1,0 +1,7 @@
+package org.ionproject.integration.application.dto
+
+data class InstitutionMetadata(
+    val name: String,
+    val acronym: String,
+    val domain: String
+)
