+++ src/main/kotlin/org/ionproject/integration/application/dto/ProgrammeMetadata.kt
@@ -1,0 +1,7 @@
+package org.ionproject.integration.application.dto
+
+data class ProgrammeMetadata(
+    val institution: InstitutionMetadata,
+    val name: String,
+    val acronym: String
+)
