+++ src/main/kotlin/org/ionproject/integration/application/job/chunkbased/FormatVerifierProcessor.kt
@@ -1,7 +1,7 @@
-package org.ionproject.integration.application.job.chunkbased.processor
+package org.ionproject.integration.application.job.chunkbased
 
+import org.ionproject.integration.application.job.ISELTimetableJob
 import org.ionproject.integration.infrastructure.IRawDataFormatChecker
-import org.ionproject.integration.job.ISELTimetableJob
 import org.ionproject.integration.domain.timetable.RawTimetableData
 import org.ionproject.integration.utils.Try
 import org.springframework.batch.item.ItemProcessor
