+++ src/main/kotlin/org/ionproject/integration/application/job/tasklet/DownloadAndCompareTasklet.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.application.job
+package org.ionproject.integration.application.job.tasklet
 
 import org.ionproject.integration.application.JobEngine.Companion.JOB_HASH_PARAMETER
 import org.ionproject.integration.application.JobEngine.Companion.REMOTE_FILE_LOCATION_PARAMETER
