+++ src/main/kotlin/org/ionproject/integration/application/job/tasklet/MappingTasklet.kt
@@ -1,7 +1,7 @@
-package org.ionproject.integration.application.job
+package org.ionproject.integration.application.job.tasklet
 
+import org.ionproject.integration.application.job.ISELTimetableJob
 import org.ionproject.integration.domain.timetable.IselTimetableTeachersBuilder
-import org.ionproject.integration.job.ISELTimetableJob
 import org.ionproject.integration.utils.orThrow
 import org.springframework.batch.core.StepContribution
 import org.springframework.batch.core.scope.context.ChunkContext
@@ -10,14 +10,14 @@ import org.springframework.batch.repeat.RepeatStatus
 import org.springframework.stereotype.Component
 
 @Component("MappingTasklet")
-class MappingTasklet(private val state: ISELTimetableJob.State) : Tasklet {
+class MappingTasklet : Tasklet {
     override fun execute(contribution: StepContribution, chunkContext: ChunkContext): RepeatStatus? {
         val builder = IselTimetableTeachersBuilder()
 
-        builder.setTimetable(state.rawTimetableData)
-        builder.setTeachers(state.rawTimetableData)
+        builder.setTimetable(ISELTimetableJob.State.rawTimetableData)
+        builder.setTeachers(ISELTimetableJob.State.rawTimetableData)
 
-        state.timetableTeachers = builder.getTimetableTeachers().orThrow()
+        ISELTimetableJob.State.timetableTeachers = builder.getTimetableTeachers().orThrow()
 
         return RepeatStatus.FINISHED
     }
