+++ src/test/kotlin/org/ionproject/integration/application/job/chunkbased/processor/FormatVerifierProcessorTest.kt
@@ -1,6 +1,7 @@
 package org.ionproject.integration.application.job.chunkbased.processor
 
 import org.ionproject.integration.IOnIntegrationApplication
+import org.ionproject.integration.application.job.chunkbased.FormatVerifierProcessor
 import org.ionproject.integration.infrastructure.exceptions.FormatCheckException
 import org.ionproject.integration.job.ISELTimetableJob
 import org.ionproject.integration.domain.timetable.RawTimetableData
