+++ src/test/kotlin/org/ionproject/integration/step/tasklet/iseltimetable/implementations/PostUploadTaskletTest.kt
@@ -6,7 +6,7 @@ import org.ionproject.integration.IOnIntegrationApplication
 import org.ionproject.integration.application.JobEngine.Companion.JOB_HASH_PARAMETER
 import org.ionproject.integration.application.JobEngine.Companion.REMOTE_FILE_LOCATION_PARAMETER
 import org.ionproject.integration.application.JobEngine.Companion.TIMESTAMP_PARAMETER
-import org.ionproject.integration.application.job.PostUploadTasklet
+import org.ionproject.integration.application.job.tasklet.PostUploadTasklet
 import org.ionproject.integration.infrastructure.HashRepositoryImpl
 import org.ionproject.integration.job.ISELTimetableJob
 import org.ionproject.integration.job.TIMETABLE_JOB_NAME
